Implementing mitosis detection
Found a mitosis detection API which uses deep learning. The API takes in a part of a cancer slide image and predicts the likelihood of the region containing mitotic figures. I implemented methods for POSTing regions of a cancer slide image to the API and collecting the predictions in a 2D array