Uploading mitosis data to mongo
When a cancer slide image is posted to the image processing server, it will calculate the mitosis probabilities and include them in the image's MongoDB document. However, I am having an error with querying the mitosis data in React. I think the fix would be to include a field mitosisData in the GraphQL schema for ImageSet.